+++ tests/PuppeteerSharp.Contrib.Tests/Extensions/ElementHandleExtensionsTests.cs
@@ -24,10 +24,10 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
             var html = await Page.QuerySelectorAsync("html");
 
             var foo = await html.QuerySelectorWithContentAsync("div", "Foo");
-            Assert.Equal("foo", foo.Id());
+            Assert.Equal("foo", await foo.IdAsync());
 
             var bar = await html.QuerySelectorWithContentAsync("div", "Ba.");
-            Assert.Equal("bar", bar.Id());
+            Assert.Equal("bar", await bar.IdAsync());
 
             var missing = await html.QuerySelectorWithContentAsync("div", "Missing");
             Assert.Null(missing);
@@ -46,10 +46,10 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
             var html = await Page.QuerySelectorAsync("html");
 
             var divs = await html.QuerySelectorAllWithContentAsync("div", "Foo");
-            Assert.Equal(new[] { "foo" }, divs.Select(x => x.Id()));
+            Assert.Equal(new[] { "foo" }, await Task.WhenAll(divs.Select(x => x.IdAsync())));
 
             divs = await html.QuerySelectorAllWithContentAsync("div", "Ba.");
-            Assert.Equal(new[] { "bar", "baz" }, divs.Select(x => x.Id()));
+            Assert.Equal(new[] { "bar", "baz" }, await Task.WhenAll(divs.Select(x => x.IdAsync())));
 
             var missing = await html.QuerySelectorAllWithContentAsync("div", "Missing");
             Assert.Empty(missing);
@@ -61,156 +61,110 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
             var tweet = await Page.QuerySelectorAsync(".tweet");
             Assert.True(tweet.Exists());
 
-            var tweetTask = Page.QuerySelectorAsync(".tweet");
-            Assert.True(tweetTask.Exists());
-
             var missing = await Page.QuerySelectorAsync(".missing");
             Assert.False(missing.Exists());
-
-            var missingTask = Page.QuerySelectorAsync(".missing");
-            Assert.False(missingTask.Exists());
         }
 
         [Fact]
-        public async Task InnerHtml_should_return_the_inner_html_of_the_element()
+        public async Task InnerHtmlAsync_should_return_the_inner_html_of_the_element()
         {
             var like = await Page.QuerySelectorAsync(".like");
             Assert.Equal("100", await like.InnerHtmlAsync());
 
-            var retweets = await Page.QuerySelectorAsync(".retweets");
-            Assert.Equal("10", retweets.InnerHtml());
-
-            Assert.Equal("10", Page.QuerySelectorAsync(".retweets").InnerHtml());
-
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.InnerHtml());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.InnerHtmlAsync());
         }
 
         [Fact]
-        public async Task OuterHtml_should_return_the_outer_html_of_the_element()
+        public async Task OuterHtmlAsync_should_return_the_outer_html_of_the_element()
         {
             var like = await Page.QuerySelectorAsync(".like");
             Assert.Equal("<div class=\"like\">100</div>", await like.OuterHtmlAsync());
 
-            var retweets = await Page.QuerySelectorAsync(".retweets");
-            Assert.Equal("<div class=\"retweets\">10</div>", retweets.OuterHtml());
-
-            Assert.Equal("<div class=\"retweets\">10</div>", Page.QuerySelectorAsync(".retweets").OuterHtml());
-
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.OuterHtml());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.OuterHtmlAsync());
         }
 
         [Fact]
-        public async Task TextContent_should_return_the_text_content_of_the_element()
+        public async Task TextContentAsync_should_return_the_text_content_of_the_element()
         {
             var like = await Page.QuerySelectorAsync(".like");
             Assert.Equal("100", await like.TextContentAsync());
 
-            var retweets = await Page.QuerySelectorAsync(".retweets");
-            Assert.Equal("10", retweets.TextContent());
-
-            Assert.Equal("10", Page.QuerySelectorAsync(".retweets").TextContent());
-
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.TextContent());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.TextContentAsync());
         }
 
         [Fact]
-        public async Task InnerText_should_return_the_rendered_text_content_of_the_element()
+        public async Task InnerTextAsync_should_return_the_rendered_text_content_of_the_element()
         {
             var like = await Page.QuerySelectorAsync(".like");
             Assert.Equal("100", await like.InnerTextAsync());
 
-            var retweets = await Page.QuerySelectorAsync(".retweets");
-            Assert.Equal("10", retweets.InnerText());
-
-            Assert.Equal("10", Page.QuerySelectorAsync(".retweets").InnerText());
-
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.InnerText());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.InnerTextAsync());
         }
 
         [Fact]
-        public async Task HasContent_should_return_true_if_element_has_the_content()
+        public async Task HasContentAsync_should_return_true_if_element_has_the_content()
         {
             var like = await Page.QuerySelectorAsync(".like");
             Assert.True(await like.HasContentAsync("100"));
 
-            var retweets = await Page.QuerySelectorAsync(".retweets");
-            Assert.True(retweets.HasContent("10"));
-
-            Assert.True(Page.QuerySelectorAsync(".retweets").HasContent("10"));
-
-            retweets = await Page.QuerySelectorAsync("div");
-            Assert.False(retweets.HasContent("20"));
+            var retweets = await Page.QuerySelectorAsync("div");
+            Assert.False(await retweets.HasContentAsync("20"));
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.HasContent(""));
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.HasContentAsync(""));
         }
 
         [Fact]
-        public async Task ClassName_should_return_the_class_of_the_element()
+        public async Task ClassNameAsync_should_return_the_class_of_the_element()
         {
             await Page.SetContentAsync("<html><body><div class='foo bar' /></body></html>");
 
             var div = await Page.QuerySelectorAsync("div");
             Assert.Equal("foo bar", await div.ClassNameAsync());
 
-            div = await Page.QuerySelectorAsync("div");
-            Assert.Equal("foo bar", div.ClassName());
-
-            Assert.Equal("foo bar", Page.QuerySelectorAsync("div").ClassName());
-
             var body = await Page.QuerySelectorAsync("body");
-            Assert.Empty(body.ClassName());
+            Assert.Empty(await body.ClassNameAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.ClassName());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.ClassNameAsync());
         }
 
         [Fact]
-        public async Task ClassList_should_return_an_array_of_classes_of_the_element()
+        public async Task ClassListAsync_should_return_an_array_of_classes_of_the_element()
         {
             await Page.SetContentAsync("<html><body><div class='foo bar' /></body></html>");
 
             var div = await Page.QuerySelectorAsync("div");
             Assert.Equal(new[] { "foo", "bar" }, await div.ClassListAsync());
 
-            div = await Page.QuerySelectorAsync("div");
-            Assert.Equal(new[] { "foo", "bar" }, div.ClassList());
-
-            Assert.Equal(new[] { "foo", "bar" }, Page.QuerySelectorAsync("div").ClassList());
-
             var body = await Page.QuerySelectorAsync("body");
-            Assert.Empty(body.ClassList());
+            Assert.Empty(await body.ClassListAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.ClassList());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.ClassListAsync());
         }
 
         [Fact]
-        public async Task HasClass_should_return_true_if_the_element_has_the_class()
+        public async Task HasClassAsync_should_return_true_if_the_element_has_the_class()
         {
             await Page.SetContentAsync("<html><body><div class='foo bar' /></body></html>");
 
             var div = await Page.QuerySelectorAsync("div");
             Assert.True(await div.HasClassAsync("foo"));
 
-            div = await Page.QuerySelectorAsync("div");
-            Assert.True(div.HasClass("bar"));
-
-            Assert.True(Page.QuerySelectorAsync("div").HasClass("foo"));
-
             var body = await Page.QuerySelectorAsync("body");
-            Assert.False(body.HasClass(""));
+            Assert.False(await body.HasClassAsync(""));
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.HasClass(""));
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.HasClassAsync(""));
         }
 
         [Fact]
-        public async Task IsVisible_should_return_true_if_the_element_is_visible()
+        public async Task IsVisibleAsync_should_return_true_if_the_element_is_visible()
         {
             await Page.SetContentAsync("<html><body><div id='foo'>Foo</div><div id='bar' style='display:none'>Bar</div><div id='baz' style='visibility:hidden'>Baz</div></body></html>");
 
@@ -218,16 +172,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
             Assert.True(await div.IsVisibleAsync());
 
             div = await Page.QuerySelectorAsync("#bar");
-            Assert.False(div.IsVisible());
+            Assert.False(await div.IsVisibleAsync());
 
-            Assert.True(Page.QuerySelectorAsync("#baz").IsVisible()); // according to the jQuery implementation
+            div = await Page.QuerySelectorAsync("#baz");
+            Assert.True(await div.IsVisibleAsync()); // according to the jQuery implementation
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsVisible());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsVisibleAsync());
         }
 
         [Fact]
-        public async Task IsHidden_should_return_false_if_the_element_is_visible()
+        public async Task IsHiddenAsync_should_return_false_if_the_element_is_visible()
         {
             await Page.SetContentAsync("<html><body><div id='foo'>Foo</div><div id='bar' style='display:none'>Bar</div><div id='baz' style='visibility:hidden'>Baz</div></body></html>");
 
@@ -235,16 +190,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
             Assert.False(await div.IsHiddenAsync());
 
             div = await Page.QuerySelectorAsync("#bar");
-            Assert.True(div.IsHidden());
+            Assert.True(await div.IsHiddenAsync());
 
-            Assert.False(Page.QuerySelectorAsync("#baz").IsHidden()); // according to the jQuery implementation
+            div = await Page.QuerySelectorAsync("#baz");
+            Assert.False(await div.IsHiddenAsync()); // according to the jQuery implementation
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsHidden());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsHiddenAsync());
         }
 
         [Fact]
-        public async Task IsSelected_should_return_true_if_the_element_is_selected()
+        public async Task IsSelectedAsync_should_return_true_if_the_element_is_selected()
         {
             await Page.SetContentAsync("<html><body><select><option id='foo'>Foo</option><option id='bar'>Bar</option><option id='baz'>Baz</option></select></body></html>");
 
@@ -253,16 +209,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
 
             await Page.SelectAsync("select", "Bar");
             option = await Page.QuerySelectorAsync("#bar");
-            Assert.True(option.IsSelected());
+            Assert.True(await option.IsSelectedAsync());
 
-            Assert.False(Page.QuerySelectorAsync("#baz").IsSelected());
+            option = await Page.QuerySelectorAsync("#baz");
+            Assert.False(await option.IsSelectedAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsSelected());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsSelectedAsync());
         }
 
         [Fact]
-        public async Task IsChecked_should_return_true_if_the_element_is_checked()
+        public async Task IsCheckedAsync_should_return_true_if_the_element_is_checked()
         {
             await Page.SetContentAsync("<html><body><input type='checkbox' id='foo' checked><input type='checkbox' id='bar'><input type='checkbox' id='baz'></body></html>");
 
@@ -271,16 +228,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
 
             await Page.ClickAsync("#bar");
             input = await Page.QuerySelectorAsync("#bar");
-            Assert.True(input.IsChecked());
+            Assert.True(await input.IsCheckedAsync());
 
-            Assert.False(Page.QuerySelectorAsync("#baz").IsChecked());
+            input = await Page.QuerySelectorAsync("#baz");
+            Assert.False(await input.IsCheckedAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsChecked());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsCheckedAsync());
         }
 
         [Fact]
-        public async Task IsDisabled_should_return_true_if_the_element_is_disabled()
+        public async Task IsDisabledAsync_should_return_true_if_the_element_is_disabled()
         {
             await Page.SetContentAsync("<html><body><input id='foo' disabled><input id='bar'><input id='baz'></body></html>");
 
@@ -289,16 +247,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
 
             await Page.EvaluateExpressionAsync("document.getElementById('bar').disabled = true");
             input = await Page.QuerySelectorAsync("#bar");
-            Assert.True(input.IsDisabled());
+            Assert.True(await input.IsDisabledAsync());
 
-            Assert.False(Page.QuerySelectorAsync("#baz").IsDisabled());
+            input = await Page.QuerySelectorAsync("#baz");
+            Assert.False(await input.IsDisabledAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsDisabled());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsDisabledAsync());
         }
 
         [Fact]
-        public async Task IsEnabled_should_return_false_if_the_element_is_disabled()
+        public async Task IsEnabledAsync_should_return_false_if_the_element_is_disabled()
         {
             await Page.SetContentAsync("<html><body><input id='foo' disabled><input id='bar'><input id='baz'></body></html>");
 
@@ -307,16 +266,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
 
             await Page.EvaluateExpressionAsync("document.getElementById('bar').disabled = true");
             input = await Page.QuerySelectorAsync("#bar");
-            Assert.False(input.IsEnabled());
+            Assert.False(await input.IsEnabledAsync());
 
-            Assert.True(Page.QuerySelectorAsync("#baz").IsEnabled());
+            input = await Page.QuerySelectorAsync("#baz");
+            Assert.True(await input.IsEnabledAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsEnabled());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsEnabledAsync());
         }
 
         [Fact]
-        public async Task IsReadOnly_should_return_true_if_the_element_is_readonly()
+        public async Task IsReadOnlyAsync_should_return_true_if_the_element_is_readonly()
         {
             await Page.SetContentAsync("<html><body><input id='foo' readonly><input id='bar'><input id='baz'></body></html>");
 
@@ -325,16 +285,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
 
             await Page.EvaluateExpressionAsync("document.getElementById('bar').readOnly = true");
             input = await Page.QuerySelectorAsync("#bar");
-            Assert.True(input.IsReadOnly());
+            Assert.True(await input.IsReadOnlyAsync());
 
-            Assert.False(Page.QuerySelectorAsync("#baz").IsReadOnly());
+            input = await Page.QuerySelectorAsync("#baz");
+            Assert.False(await input.IsReadOnlyAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsReadOnly());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsReadOnlyAsync());
         }
 
         [Fact]
-        public async Task IsRequired_should_return_true_if_the_element_is_required()
+        public async Task IsRequiredAsync_should_return_true_if_the_element_is_required()
         {
             await Page.SetContentAsync("<html><body><input id='foo' required><input id='bar'><input id='baz'></body></html>");
 
@@ -343,16 +304,17 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
 
             await Page.EvaluateExpressionAsync("document.getElementById('bar').required = true");
             input = await Page.QuerySelectorAsync("#bar");
-            Assert.True(input.IsRequired());
+            Assert.True(await input.IsRequiredAsync());
 
-            Assert.False(Page.QuerySelectorAsync("#baz").IsRequired());
+            input = await Page.QuerySelectorAsync("#baz");
+            Assert.False(await input.IsRequiredAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.IsRequired());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.IsRequiredAsync());
         }
 
         [Fact]
-        public async Task HasFocus_should_return_true_if_the_element_has_focus()
+        public async Task HasFocusAsync_should_return_true_if_the_element_has_focus()
         {
             await Page.SetContentAsync("<html><body><input id='foo' autofocus><input id='bar'><input id='baz'></body></html>", new NavigationOptions { WaitUntil = new[] { WaitUntilNavigation.Networkidle0 } });
 
@@ -361,12 +323,13 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
 
             await Page.FocusAsync("#bar");
             input = await Page.QuerySelectorAsync("#bar");
-            Assert.True(input.HasFocus());
+            Assert.True(await input.HasFocusAsync());
 
-            Assert.False(Page.QuerySelectorAsync("#baz").HasFocus());
+            input = await Page.QuerySelectorAsync("#baz");
+            Assert.False(await input.HasFocusAsync());
 
             var missing = await Page.QuerySelectorAsync(".missing");
-            Assert.Throws<ArgumentNullException>(() => missing.HasFocus());
+            await Assert.ThrowsAsync<ArgumentNullException>(() => missing.HasFocusAsync());
         }
     }
 }
