+++ tests/PuppeteerSharp.Contrib.Tests/Extensions/PageExtensionsTests.cs
@@ -19,10 +19,10 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
         public async Task QuerySelectorWithContentAsync_should_return_the_first_element_that_match_the_selector_and_has_the_content()
         {
             var foo = await Page.QuerySelectorWithContentAsync("div", "Foo");
-            Assert.Equal("foo", foo.Id());
+            Assert.Equal("foo", await foo.IdAsync());
 
             var bar = await Page.QuerySelectorWithContentAsync("div", "Ba.");
-            Assert.Equal("bar", bar.Id());
+            Assert.Equal("bar", await bar.IdAsync());
 
             var missing = await Page.QuerySelectorWithContentAsync("div", "Missing");
             Assert.Null(missing);
@@ -32,10 +32,10 @@ namespace PuppeteerSharp.Contrib.Tests.Extensions
         public async Task QuerySelectorAllWithContentAsync_should_return_all_elements_that_match_the_selector_and_has_the_content()
         {
             var divs = await Page.QuerySelectorAllWithContentAsync("div", "Foo");
-            Assert.Equal(new[] { "foo" }, divs.Select(x => x.Id()));
+            Assert.Equal(new[] { "foo" }, await Task.WhenAll(divs.Select(x => x.IdAsync())));
 
             divs = await Page.QuerySelectorAllWithContentAsync("div", "Ba.");
-            Assert.Equal(new[] { "bar", "baz" }, divs.Select(x => x.Id()));
+            Assert.Equal(new[] { "bar", "baz" }, await Task.WhenAll(divs.Select(x => x.IdAsync())));
 
             var missing = await Page.QuerySelectorAllWithContentAsync("div", "Missing");
             Assert.Empty(missing);
